feat(standard-plugins/nvim.undotree): add lazy loading
This makes it more convenient to use the new standard plugin by lazy
loading it automatically on command usage.

It is the very first standard plugin that gets used in the
configuration. Therefore, the new package was created. The naming of
"standard-plugins" comes from the `:help usr_05.txt` document. Hence
also another exception for the Lua file name convention.