feat(colorscheme): establish set of custom VcsStatus highlight groups
These are meant to be complementary to the native highlight groups like
`DiffAdd` or `DiffChange`. But while the latter are meant to highlight
the buffer hunks in a diff, the added one are meant for version control
related UI elements. See the changed files for some user cases.

The naming of the statuses are strictly taken from the `git-status` man
page.